requirements-testing.txt
pytest==3.0.7
pytest-django==3.1.2
flake8==3.3.0
coverage==4.3.4
factory-boy==2.8.1
